style: update

+++ src/utils.hpp
@@ -291,7 +291,7 @@ static inline bool is_base64(unsigned char c) {
     return (isalnum(c) || (c == '+') || (c == '/'));
 }
 
-std::string base64_encode(unsigned char const *bytes_to_encode, unsigned int in_len) {
+std::string base64_encode(unsigned char const *bytes_to_encode, unsigned long in_len) {
     std::string ret;
     int i = 0;
     int j = 0;
